split http request handlers
